refactor setup.py
